+++ src/components/Admin/AdminDataViews/AdminDataViews.tsx
@@ -1,0 +1,272 @@
+import React, { useState } from 'react'
+import { useRouter } from '../../../utils/hooks/useRouter'
+import {
+  Header,
+  Button,
+  Dropdown,
+  Checkbox,
+  Icon,
+  DropdownItemProps,
+  DropdownProps,
+} from 'semantic-ui-react'
+import { TranslateMethod, useLanguageProvider } from '../../../contexts/Localisation'
+import usePageTitle from '../../../utils/hooks/usePageTitle'
+import useConfirmationModal from '../../../utils/hooks/useConfirmationModal'
+import {
+  DataTable,
+  DataView,
+  GetDataTablesQuery,
+  useGetDataTablesQuery,
+} from '../../../utils/generated/graphql'
+import { toCamelCase } from '../../../LookupTable/utils'
+import { JsonEditor } from '../JsonEditor'
+import { pickBy } from 'lodash'
+import { nanoid } from 'nanoid'
+import { useAdminDataViewConfig } from './DatabaseHooks'
+
+export const AdminDataViews: React.FC = () => {
+  const { t } = useLanguageProvider()
+  usePageTitle(t('PAGE_TITLE_DATA_VIEW'))
+  const [includeLookupTables, setIncludeLookupTables] = useState(false)
+
+  const { query, updateQuery, setQuery } = useRouter()
+
+  const { data, loading } = useGetDataTablesQuery()
+
+  const selectedTable = query.selectedTable
+
+  return (
+    <div id="data-view-config-panel" className="flex-column" style={{ gap: 15 }}>
+      <Header>{t('DATA_VIEW_CONFIG_HEADER')}</Header>
+      <div className="flex-row-space-between-center">
+        <Dropdown
+          selection
+          clearable
+          placeholder={t('DATA_VIEW_CONFIG_SELECT_TABLE')}
+          loading={loading}
+          value={selectedTable}
+          options={getDataTableOptions(data, includeLookupTables, t)}
+          onChange={(_, { value }) => {
+            if (!value) setQuery({})
+            else updateQuery({ selectedTable: value })
+          }}
+          style={{ minWidth: 300 }}
+        />
+        <Checkbox
+          checked={includeLookupTables}
+          onChange={() => setIncludeLookupTables(!includeLookupTables)}
+          label={t('DATA_VIEW_CONFIG_INCLUDE_LOOKUP')}
+        />
+      </div>
+      {selectedTable && <DataViewEditor tableName={selectedTable} />}
+      {/* <ColumnDefinitionEditor /> */}
+    </div>
+  )
+}
+
+interface DataViewEditorProps {
+  tableName: string
+}
+
+const DataViewEditor: React.FC<DataViewEditorProps> = ({ tableName }) => {
+  const { t } = useLanguageProvider()
+  const { updateQuery } = useRouter()
+  const { ConfirmModal, showModal: showConfirmation } = useConfirmationModal({
+    type: 'warning',
+    confirmText: t('BUTTON_CONFIRM'),
+  })
+
+  const {
+    selectedDataView,
+    dataViews,
+    dataViewObject,
+    loading,
+    updateDataView,
+    deleteDataView,
+    addDataView,
+    isSaving,
+    isDeleting,
+    isAdding,
+  } = useAdminDataViewConfig(tableName)
+
+  return (
+    <>
+      <ConfirmModal />
+      <DataViewDisplay
+        title="Data Views"
+        placeholder={t('DATA_VIEW_CONFIG_SELECT_VIEW')}
+        loading={loading}
+        dropdownValue={selectedDataView}
+        options={getDataViewOptions(dataViews)}
+        onChange={(_, { value }) => {
+          if (dataViews) updateQuery({ dataView: value })
+        }}
+        data={
+          dataViewObject
+            ? pickBy(
+                dataViewObject,
+                (_, key) =>
+                  !['__typename', 'id']
+                    // These two properties shouldn't show up in the editor,
+                    // as we don't want them being touched
+                    .includes(key)
+              )
+            : undefined
+        }
+        dataName={dataViewObject?.identifier as string}
+        onSave={(data) => {
+          showConfirmation({
+            title: t('DATA_VIEW_CONFIG_SAVE_WARNING'),
+            message: t('DATA_VIEW_CONFIG_SAVE_MESSAGE'),
+            onConfirm: () =>
+              updateDataView({ variables: { id: dataViewObject?.id as number, patch: data } }),
+          })
+        }}
+        isSaving={isSaving}
+        onDelete={() =>
+          showConfirmation({
+            title: t('DATA_VIEW_CONFIG_DELETE_WARNING'),
+            message: t('DATA_VIEW_CONFIG_DELETE_MESSAGE'),
+            onConfirm: () => deleteDataView({ variables: { id: dataViewObject?.id as number } }),
+          })
+        }
+        isDeleting={isDeleting}
+        onAdd={() => {
+          addDataView({
+            variables: {
+              identifier: `${tableName}_${nanoid(8)}`,
+              tableName,
+              code: '__CODE__',
+              detailViewHeaderColumn: '__HEADER_COL__',
+            },
+          })
+        }}
+        isAdding={isAdding}
+      />
+    </>
+  )
+}
+
+interface DataViewDisplayProps {
+  title: string
+  placeholder: string
+  loading: boolean
+  dropdownValue: string
+  options: DropdownItemProps[]
+  onChange: (_: unknown, value: DropdownProps) => void
+  data: object | undefined
+  dataName: string
+  onSave: (data: object) => void
+  isSaving: boolean
+  onDelete: () => void
+  isDeleting: boolean
+  onAdd: () => void
+  isAdding: boolean
+}
+
+const DataViewDisplay: React.FC<DataViewDisplayProps> = ({
+  title,
+  placeholder,
+  loading,
+  dropdownValue,
+  options,
+  onChange,
+  data,
+  dataName,
+  onSave,
+  isSaving,
+  onDelete,
+  isDeleting,
+  onAdd,
+  isAdding,
+}) => {
+  const { t } = useLanguageProvider()
+  const { ConfirmModal } = useConfirmationModal({
+    type: 'warning',
+    confirmText: t('BUTTON_CONFIRM'),
+  })
+
+  return (
+    <div>
+      <Header as="h2">{title}</Header>
+      <ConfirmModal />
+      <div className="flex-row-space-between">
+        <Dropdown
+          selection
+          clearable
+          placeholder={placeholder}
+          loading={loading}
+          value={dropdownValue}
+          options={options}
+          onChange={onChange}
+          style={{ minWidth: 300 }}
+        />
+        <div>
+          <Button
+            primary
+            inverted
+            disabled={!data}
+            loading={isDeleting}
+            icon={<Icon name="trash alternate outline" size="small" />}
+            content={t('DATA_VIEW_CONFIG_DELETE_BUTTON')}
+            onClick={onDelete}
+          />
+          <Button
+            primary
+            inverted
+            loading={isAdding}
+            icon={<Icon name="plus" size="tiny" color="blue" />}
+            content={t('DATA_VIEW_CONFIG_ADD_BUTTON')}
+            onClick={onAdd}
+          />
+        </div>
+      </div>
+      {data && (
+        <JsonEditor
+          data={data}
+          onSave={onSave}
+          isSaving={isSaving}
+          name={dataName}
+          collapsed={1}
+          displayArrayKey={false}
+          quotesOnKeys={false}
+          displayDataTypes={true}
+          style={{ padding: '10px' }}
+        />
+      )}
+    </div>
+  )
+}
+
+const getDataTableOptions = (
+  data: GetDataTablesQuery | undefined,
+  includeLookupTables: boolean,
+  t: TranslateMethod
+) => {
+  if (!data) return []
+
+  return (data.dataTables?.nodes as DataTable[])
+    .filter(({ isLookupTable }) => (includeLookupTables ? true : !isLookupTable))
+    .map(({ id, tableName, isLookupTable }) => {
+      const table = toCamelCase(tableName)
+      return {
+        key: `${table}_${id}`,
+        text: `${table}${isLookupTable ? ` (${t('LOOKUP_TABLE_TITLE')})` : ''}`,
+        value: table,
+      }
+    })
+}
+
+const getDataViewOptions = (dataViews: DataView[] | undefined) => {
+  if (!dataViews) return []
+
+  return dataViews.map((dataView) => {
+    const { id, identifier, code } = dataView
+    return {
+      key: `${code}_${id}`,
+      text: identifier,
+      value: identifier,
+      data: dataView,
+    }
+  })
+}
