+++ src/utils/defaultLanguageStrings.ts
@@ -135,6 +135,18 @@ export default {
   DATA_VIEW_CONFIG_ADD_MESSAGE: 'Please give it a meaningful unique identifier',
   DATA_VIEW_CONFIG_ADD_BUTTON: 'Add new',
   DATA_VIEW_CONFIG_DELETE_BUTTON: 'Delete',
+  DATA_VIEW_CONFIG_SAVE_COL_WARNING: 'Save column definition?',
+  DATA_VIEW_CONFIG_SAVE_COL_MESSAGE: 'This will overwrite the current column configuration',
+  DATA_VIEW_CONFIG_UPDATE_COL_PROBLEM: 'Problem saving column definition!',
+  DATA_VIEW_CONFIG_COL_SAVED: 'Column definition saved',
+  DATA_VIEW_CONFIG_DELETE_COL_WARNING: 'Delete column definition?',
+  DATA_VIEW_CONFIG_DELETE_COL_MESSAGE:
+    'This will irreversibly delete the column definition from the system',
+  DATA_VIEW_CONFIG_DELETE_COL_PROBLEM: 'Problem deleting column definition!',
+  DATA_VIEW_CONFIG_COL_DELETED: 'Column definition deleted',
+  DATA_VIEW_CONFIG_ADD_COL_PROBLEM: 'Problem creating column definition!',
+  DATA_VIEW_CONFIG_COL_ADDED: 'Column definition created',
+  DATA_VIEW_CONFIG_ADD_COL_MESSAGE: 'UPDATE THIS MESSAGE',
   DEFAULT_TEMPLATE_CATEGORY_TITLE: 'General',
   DEFAULT_TEMPLATE_CATEGORY_ICON: 'globe',
   ERROR_APPLICATION_CREATE: 'Problem loading application creation page',
