+++ src/utils/graphql/mutations/dataViews/createColumnDefinition.mutation.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client'
 
 export default gql`
-  mutation createDataViewColumnDefinition($tableName: String!, $columnName: String!) {
+  mutation createColumnDefinition($tableName: String!, $columnName: String!) {
     createDataViewColumnDefinition(
       input: { dataViewColumnDefinition: { tableName: $tableName, columnName: $columnName } }
     ) {
