+++ src/utils/graphql/mutations/dataViews/deleteColumnDefinition.mutation.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client'
 
 export default gql`
-  mutation deleteDataViewColumnDefinition($id: Int!) {
+  mutation deleteColumnDefinition($id: Int!) {
     deleteDataViewColumnDefinition(input: { id: $id }) {
       dataViewColumnDefinition {
         id
