+++ src/utils/graphql/mutations/dataViews/updateColumnDefinition.mutation.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client'
 
 export default gql`
-  mutation updateDataViewColumnDefinition($id: Int!, $patch: DataViewColumnDefinitionPatch!) {
+  mutation updateColumnDefinition($id: Int!, $patch: DataViewColumnDefinitionPatch!) {
     updateDataViewColumnDefinition(input: { patch: $patch, id: $id }) {
       dataViewColumnDefinition {
         ...dataViewColumnDefinitionFragment
